requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0

# Data Science & Numerical Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Computer Vision
opencv-python>=4.7.0
albumentations>=1.3.0
Pillow>=9.5.0

# Metrics & Evaluation
torchmetrics>=0.11.0
scikit-learn>=1.2.0
pytorch-grad-cam>=1.4.0

# Web Application
streamlit>=1.22.0
gradio>=3.28.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Experiment Tracking (Optional)
wandb>=0.15.0
tensorboard>=2.12.0

# API Development (Optional)
flask>=2.3.0
flask-cors>=3.0.10
fastapi>=0.95.0
uvicorn>=0.21.0
